separate theme provider
