+++ packages/iris-grid/src/index.ts
@@ -25,6 +25,7 @@ export { default as IrisGridModelFactory } from './IrisGridModelFactory';
 export { createDefaultIrisGridTheme } from './IrisGridTheme';
 export type { IrisGridThemeType } from './IrisGridTheme';
 export * from './IrisGridContextProvider';
+export * from './IrisGridThemeProvider';
 export { default as IrisGridTestUtils } from './IrisGridTestUtils';
 export { default as IrisGridUtils } from './IrisGridUtils';
 export * from './IrisGridUtils';
